+++ pytsite/console/__init__.py
@@ -12,4 +12,5 @@ __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
+
 _lang.register_package(__name__)